Split mewbot.component into core and API sections
The parts that make up the registry system have been moved into
mewbot.api.registry, whereas the definition for the YAML configuration
has been moved into mewbot.config.
The base 'Component' class has become API v1 specific, as all other
checks are now based on interface/protocol compatibility.
